+++ zetaclient/evm/evm_client.go
@@ -40,6 +40,7 @@ import (
 	crosschaintypes "github.com/zeta-chain/zetacore/x/crosschain/types"
 	observertypes "github.com/zeta-chain/zetacore/x/observer/types"
 	clientcommon "github.com/zeta-chain/zetacore/zetaclient/common"
+	"github.com/zeta-chain/zetacore/zetaclient/compliance"
 	"github.com/zeta-chain/zetacore/zetaclient/config"
 	clienttypes "github.com/zeta-chain/zetacore/zetaclient/types"
 	"gorm.io/driver/sqlite"
@@ -330,7 +331,7 @@ func (ob *ChainClient) IsSendOutTxProcessed(cctx *crosschaintypes.CrossChainTx, 
 	logger = logger.With().Str("sendID", sendID).Logger()
 
 	// compliance check, special handling the cancelled cctx
-	if clientcommon.IsCctxRestricted(cctx) {
+	if compliance.IsCctxRestricted(cctx) {
 		recvStatus := common.ReceiveStatus_Failed
 		if receipt.Status == 1 {
 			recvStatus = common.ReceiveStatus_Success
@@ -907,10 +908,7 @@ func (ob *ChainClient) postBlockHeader(tip uint64) error {
 
 func (ob *ChainClient) observeInTX(sampledLogger zerolog.Logger) error {
 	// make sure inbound TXS / Send is enabled by the protocol
-	flags, err := ob.zetaClient.GetCrosschainFlags()
-	if err != nil {
-		return err
-	}
+	flags := ob.coreContext.GetCrossChainFlags()
 	if !flags.IsInboundEnabled {
 		return errors.New("inbound TXS / Send has been disabled by the protocol")
 	}