+++ zetaclient/evm/evm_signer.go
@@ -25,6 +25,8 @@ import (
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
 	observertypes "github.com/zeta-chain/zetacore/x/observer/types"
 	clientcommon "github.com/zeta-chain/zetacore/zetaclient/common"
+	"github.com/zeta-chain/zetacore/zetaclient/compliance"
+	corecontext "github.com/zeta-chain/zetacore/zetaclient/core_context"
 	"github.com/zeta-chain/zetacore/zetaclient/interfaces"
 	"github.com/zeta-chain/zetacore/zetaclient/metrics"
 	"github.com/zeta-chain/zetacore/zetaclient/outtxprocessor"
@@ -34,12 +36,13 @@ import (
 
 // Signer deals with the signing EVM transactions and implements the ChainSigner interface
 type Signer struct {
-	client    interfaces.EVMRPCClient
-	chain     *common.Chain
-	tssSigner interfaces.TSSSigner
-	ethSigner ethtypes.Signer
-	logger    clientcommon.ClientLogger
-	ts        *metrics.TelemetryServer
+	client      interfaces.EVMRPCClient
+	chain       *common.Chain
+	tssSigner   interfaces.TSSSigner
+	ethSigner   ethtypes.Signer
+	logger      clientcommon.ClientLogger
+	ts          *metrics.TelemetryServer
+	coreContext *corecontext.ZetaCoreContext
 
 	// mu protects below fields from concurrent access
 	mu                     *sync.Mutex
@@ -60,6 +63,7 @@ func NewEVMSigner(
 	erc20CustodyABI string,
 	zetaConnectorAddress ethcommon.Address,
 	erc20CustodyAddress ethcommon.Address,
+	coreContext *corecontext.ZetaCoreContext,
 	loggers clientcommon.ClientLogger,
 	ts *metrics.TelemetryServer,
 ) (*Signer, error) {
@@ -85,6 +89,7 @@ func NewEVMSigner(
 		erc20CustodyABI:      custodyABI,
 		zetaConnectorAddress: zetaConnectorAddress,
 		er20CustodyAddress:   erc20CustodyAddress,
+		coreContext:          coreContext,
 		logger: clientcommon.ClientLogger{
 			Std:        loggers.Std.With().Str("chain", chain.ChainName.String()).Str("module", "EVMSigner").Logger(),
 			Compliance: loggers.Compliance,
@@ -341,16 +346,12 @@ func (signer *Signer) TryProcessOutTx(
 	toChain := common.GetChainFromChainID(txData.toChainID.Int64())
 
 	// Get cross-chain flags
-	crossChainflags, err := zetaBridge.GetCrosschainFlags()
-	if err != nil {
-		logger.Err(err).Msg("couldn't retrieve crosschain flags from core")
-		return
-	}
+	crossChainflags := signer.coreContext.GetCrossChainFlags()
 
 	var tx *ethtypes.Transaction
 	// compliance check goes first
-	if clientcommon.IsCctxRestricted(cctx) {
-		clientcommon.PrintComplianceLog(logger, signer.logger.Compliance,
+	if compliance.IsCctxRestricted(cctx) {
+		compliance.PrintComplianceLog(logger, signer.logger.Compliance,
 			true, evmClient.chain.ChainId, cctx.Index, cctx.InboundTxParams.Sender, txData.to.Hex(), cctx.InboundTxParams.CoinType.String())
 		tx, err = signer.SignCancelTx(txData.nonce, txData.gasPrice, height) // cancel the tx
 		if err != nil {